Changed book cover art.
The previous cover was an image from the Guild of Writers, and nobody there would get back to me about whether or not I had permission to use it. Since this project is in the public domain, I have switched to the Mystcraft version of the "falling man" with XCompWiz's permission.